atmk_system-master/requirements.txt
beautifulsoup4==4.10.0
Django==3.2.5
gensim==4.1.2
h5py==3.6.0
jieba==0.42.1
keras==2.7.0
numpy==1.21.1
PyYAML==6.0
tensorflow==2.7.0
torch==1.10.2
